Initial commit. Added mosquitto authorization plugin.
Mosquitto's default password and acl management is too simple and
cumbersome to securely manage a project as complex and dynamic as
an IoT automation device network. So, an SQLite3-based authorization
plugin has been implemented, capable of interacting with a central
database which might be updated by an external process.

The plugin generates the appropriate database file if it is not found,
partitions the channels giving access only to a unique subtopic to
each device, gives access to a lobby (also partitioned per device)
to any device holding a secret guest password and considers the
presence of a superuser server, who has access to every channel.