Added web server to provide static data.
Any "public" data will be posted on this web server, e.g. new WiFi
credentials, firmware updates. The only requirement is to be able
to connect to the network (e.g. having the old WiFi password).

The service uses SSL to secure communications, but do note there
is sensitive information in this server, i.e. the firmware binary
will contain the guest secret, and there's no client verification
because certificates are too memory heavy for most devices. Hence,
having the WiFi password means having access to the device network.
Choose a good password!

Such an attacker could gain the guest secret to pose as a new device
but would not be able to fool devices into executing commands, since
server certificate verification will be performed on the devices.
Also, any new device must be vetted by the final user to be granted
any relevant permissions, so the surface area for an intrusion is
pretty much limited to Denial-of-Service attacks. Since the attacker
must be able to access the physical WiFi network, at that point it may
very well jam the WiFi signal for better results.