Implemented device master orchestrator.
An early implementation of the main server program as a python script.
It configures and runs the Mosquitto broker on the background and
processes the messages from the final user to the devices and
vice versa.

The public interface consist of a CLI command processor through the
standard input/output.
Currently, the commands include adding/removing devices to the network,
synchronizing their time, sending PING requests, asking their status,
executing device-specific commands and scheduling commands for a later
time (either at a specific date or in a recurrent fashion).

Naturally this interface is not meant for humans to use directly,
but as a REST backend for a graphical web server.

Note that appropriate SSL keys must be generated to use the server.
Modify the credentials.json template to reflect any changes.