Added Sonoff firmware.
An Arduino-ESP8266 based firmware has been added. It uses PubSubClient
to connect to the MQTT network. It can turn the Sonoff switch on
and off, toggle it and schedule commands for a later date (either
a fixed date or a recurrent one).

It uses the internal ESP8266 millis() clock to keep track of time
so scheduled events do not rely on the server. That way in case of
a network failure, the device can still fulfill its duty. Note
however, that the internal clock of the ESP8266 is not particularly
precise (drift of around a second per hour), so the server should
try to synchronize on a regular basis. Nevertheless, the drift may
not be a big deal for many applications such as turning the lights on.

The device can also schedule a fuzzy event, which will trigger around
the set time, but with -+8 minutes spread. This may be useful when
turning a lot of devices at the same time. This way, the electrical
power surge is not so pronounced, the network does not get so
loaded if the devices access an online resource and it doesn't look
visually ackward and suspicious (e.g. when all the lights in the house
turn on at the same time).