Status-dependent command validity.
The validcommand() function only received two arguments: the device
type and the command. This is fine but many sensor usually have
certain commands that only make sense in certain configurations,
so the status has been added as an argument for the validity check.

Also, a general code lint was performed.
* Changed function names to be consistent with each other.
* Added missing documentation.
* Other minor aesthetic changes.