Request device's schedule.
So far, the device's internal schedule, one of its most important
pieces of state was inaccesible to the host. This meant that any
corruption would go unnoticed, be it the schedule message never
reached the device, it suffered a hard reset or any other problem.

Now, the new askschedule call gives the host the device's schedule.
It then can compare it to its own database and fix any inconsistencies
(this fix step is done automatically after asking for the schedule
if any differences were observed).

The firmware has been updated to respond to the request.

Also, two major refactorings:
* The event management has been centralized in a class, to
  simplify most interfaces (instead of calling several every function
  with the same arguments over and over again).
* The two main routines were very long. One has been divided into
  subroutines and in the other one a big switch-like if statement
  has been changed into a lookup table, reducing a lot of
  boilerplate details.